requirements.txt
pynacl  # needs libffi-dev
argon2_cffi
Flask-JSONRPC
requests
